Move utils to jest folder
