+++ present_ideas/present_ideas/static/components/__tests__/LoadingIndicator.test.ts
@@ -1,4 +1,4 @@
-import { render } from "../../pages/__tests__/utils";
+import { render } from "../../../../../jest/utils";
 import LoadingIndicator from "../LoadingIndicator.vue";
 
 describe("LoadingIndicator", () => {
