+++ present_ideas/present_ideas/static/components/__tests__/PeoplePicker.test.ts
@@ -1,4 +1,4 @@
-import { render } from "../../pages/__tests__/utils";
+import { render } from "../../../../../jest/utils";
 import PeoplePicker from "../PeoplePicker.vue";
 
 describe("PeoplePicker", () => {
