+++ present_ideas/present_ideas/static/components/__tests__/PresentIdeaFormCard.test.ts
@@ -1,4 +1,4 @@
-import { render } from "../../pages/__tests__/utils";
+import { render } from "../../../../../jest/utils";
 import PresentIdeaFormCard from "../PresentIdeaFormCard.vue";
 
 describe("PresentIdeaFormCard", () => {
