+++ present_ideas/present_ideas/static/layout/__tests__/Layout.test.ts
@@ -1,4 +1,4 @@
-import { render } from "../../pages/__tests__/utils";
+import { render } from "../../../../../jest/utils";
 import Layout from "../Layout.vue";
 
 describe("Layout", () => {
