+++ present_ideas/present_ideas/static/pages/__tests__/LandingPage.test.ts
@@ -1,5 +1,5 @@
+import { render } from "../../../../../jest/utils";
 import LandingPage from "../LandingPage.vue";
-import { render } from "./utils";
 
 describe("LandingPage", () => {
   it("matches snapshot", () => {
