+++ present_ideas/present_ideas/static/store/__tests__/store.test.ts
@@ -1,6 +1,6 @@
 import { rest } from "msw";
 import { server } from "../../../../../jest/mock_server";
-import { getVueAndStore } from "../../pages/__tests__/utils";
+import { getVueAndStore } from "../../../../../jest/utils";
 
 describe("getters", () => {
   it("returns an empty list for people initially", () => {
